Rename gui
Cambiato nome al file `gui.py` in `GUI.py`
